Fixed review issues pointed out by Ms2ger
