+++ src/components/util/util.rs
@@ -22,7 +22,6 @@ extern crate sync;
 extern crate std_time = "time";
 extern crate std_url = "url";
 
-pub mod attr;
 pub mod cache;
 pub mod concurrentmap;
 pub mod cowarc;
